addressing rchk, valgrind and ubsan issues [no ci]

+++ src/conversion.c
@@ -233,7 +233,7 @@ SEXP array_to_nr_(SEXP arr_, SEXP dst_) {
     for (int col = 0; col < width; col++) {
       for (int row = 0; row < height; row++) {
         *(nr + row * width + col) = 
-          (0xff << 24) +
+          ((unsigned int)0xff << 24) +
           ((uint8_t)(*(b + col * height + row) * 255) << 16) + 
           ((uint8_t)(*(g + col * height + row) * 255) <<  8) +
           ((uint8_t)(*(r + col * height + row) * 255)      );
